# shared/contracts/openapi/cloudrun-api.yaml
openapi: 3.0.3
info:
  title: Anti-Scam Guardian API
  version: 0.1.0
paths:
  /v1/signals/analyze:
    post:
      summary: Analyze signal from mobile client
      operationId: analyzeSignal
      tags:
        - signals
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceType]
              properties:
                sourceType:
                  type: string
                  enum: [SMS, URL, CALL]
                text:
                  type: string
                callSessionId:
                  type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          description: Risk result
          content:
            application/json:
              schema:
                type: object
                properties:
                  riskScore:
                    type: integer
                  voiceScore:
                    type: integer
                  textScore:
                    type: integer
                  entityScore:
                    type: integer
                  explanation:
                    type: string
                  cacheHit:
                    type: boolean

  /v1/feedback:
    post:
      summary: Submit Scam/Safe/Not sure label
      operationId: submitFeedback
      tags:
        - feedback
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../event-schema/feedback-event.json"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT